+++ src/dataLayer/fetchData.ts
@@ -23,7 +23,7 @@ export const getTemplate = async (type: string): Promise<Template> => {
   return await postData(path, type).then((data) => data);
 };
 
-export const getNewObject = async (type: string): Promise<DataModal> => {
+export const getNewObject = async (type: string): Promise<any> => {
   const template = await getTemplate(type);
   const object = {
     template,
@@ -32,7 +32,6 @@ export const getNewObject = async (type: string): Promise<DataModal> => {
       name: '',
       type: type,
       relationships: [],
-      _id: '',
     },
   };