+++ src/edit/EditObject.tsx
@@ -26,40 +26,44 @@ const EditObject: FunctionComponent = () => {
   const urlID = window.location.pathname.split('/').pop();
   const history = useHistory();
 
-  const [object, setObject] = useState<DataModal | any>();
+  const [object, setObject] = useState<any>();
+  const [template, setTemplate] = useState<any>()
   const [type, setType] = useState<string>('Activity');
 
   const handleChange = async (value: string) => {
-    const newType = value;
-    setType(newType);
-    const template = await getTemplate(newType);
-    // have to update the relationships we can add
-
-    const tempObj = { ...object };
-    tempObj.data.type = newType;
-    tempObj.Template = template;
-    setObject(tempObj);
+    // const newType = value;
+    // setType(newType);
+    // const template = await getTemplate(newType);
+    // // have to update the relationships we can add
+    //
+    // const tempObj = { ...object };
+    // tempObj.data.type = newType;
+    // tempObj.Template = template;
+    // setObject(tempObj);
   };
 
   useEffect(() => {
     if (urlID) {
       const id = parseInt(urlID);
-      if (id === 0) {
-        (async () => {
-          const objectData = await getNewObject(type);
-          setObject(objectData);
-          setType(objectData.data.type);
-        })();
-        return;
-      }
+      // if (id === 0) {
+      //   (async () => {
+      //     const objectData = await getNewObject(type);
+      //     setObject(objectData);
+      //     setType(objectData.data.type);
+      //   })();
+      //   return;
+      // }
 
       (async () => {
-        setObject(await getObjectData(id));
+        console.log(id);
+        const { template, data } = await getObjectData(id);
+        setObject(data);
+        setTemplate(template);
       })();
     }
   }, [urlID]);
 
-  if (!object) {
+  if (!object || !template) {
     return <Loading />;
   }
 
@@ -70,93 +74,91 @@ const EditObject: FunctionComponent = () => {
     objectKey: string,
     id: number // leaving for now in case need it later
   ) {
-    console.log(id);
-    const newObject = { ...object };
-    newObject.data[objectKey] = value;
-    setObject(newObject);
+  //   const newObject = { ...object };
+  //   newObject.data[objectKey] = value;
+  //   setObject(newObject);
   }
 
   function removeRelationship(objectKey: string, removedObject: Relationship) {
-    const tempObject = { ...object };
-    const relationships = tempObject.data[objectKey].filter(
-      (relationshipObject: Relationship) =>
-        relationshipObject.id !== removedObject.id
-    );
-    const { id, type, name, _id } = tempObject.data;
-
-    // dont really like mutation but will work for now
-    tempObject.data = { id, type, name, relationships, _id };
-    setObject(tempObject);
+  //   const tempObject = { ...object };
+  //   const relationships = tempObject.data[objectKey].filter(
+  //     (relationshipObject: Relationship) =>
+  //       relationshipObject.id !== removedObject.id
+  //   );
+  //   const { id, type, name, _id } = tempObject.data;
+  //
+  //   // dont really like mutation but will work for now
+  //   tempObject.data = { id, type, name, relationships, _id };
+  //   setObject(tempObject);
   }
 
   // still need to add logic on server side
   // to save new relationship object
   function addRelationship(type, objID = 0, name = '', _id = '') {
-    const tempObject = { ...object };
-    console.log(_id);
+  //   const tempObject = { ...object };
+  //
+  //   // only do this if its not new object
+  //   if (objID !== 0) {
+  //     const isDuplicate = tempObject.data.relationships.find(
+  //       ({ id }) => id === objID
+  //     );
+  //
+  //     // no need add duplicate relationships
+  //     if (isDuplicate) {
+  //       alert('No need to duplicate');
+  //       // need to clear box after this
+  //       return;
+  //     }
+  //
+  //     // now that we know what relationship user wanted
+  //     // remove the last relationship object
+  //     // it should always be 0
+  //     tempObject.data.relationships.pop();
+  //   }
 
-    // only do this if its not new object
-    if (objID !== 0) {
-      const isDuplicate = tempObject.data.relationships.find(
-        ({ id }) => id === objID
-      );
-
-      // no need add duplicate relationships
-      if (isDuplicate) {
-        alert('No need to duplicate');
-        // need to clear box after this
-        return;
-      }
-
-      // now that we know what relationship user wanted
-      // remove the last relationship object
-      // it should always be 0
-      tempObject.data.relationships.pop();
-    }
-
-    const newRelationship = {
-      id: objID,
-      type,
-      name,
-    };
-
-    tempObject.data.relationships.push(newRelationship);
-    setObject(tempObject);
+  //   const newRelationship = {
+  //     id: objID,
+  //     type,
+  //     name,
+  //   };
+  //
+  //   tempObject.data.relationships.push(newRelationship);
+  //   setObject(tempObject);
   }
 
   async function saveObject() {
-    if (object.data.name === '') {
-      // maybe a modal here?
-      alert('Need a name');
-      return;
-    }
-
-    // need error handling
-    const obj = await updateObject(object);
-    history.push(`/view/${obj.data.id}`);
+  //   if (object.data.name === '') {
+  //     // maybe a modal here?
+  //     alert('Need a name');
+  //     return;
+  //   }
+  //
+  //   // need error handling
+  //   const obj = await updateObject(object);
+  //   history.push(`/view/${obj.data.id}`);
   }
 
   async function deleteObject() {
-    const status = await deleteObjectById(object.data._id);
-
-    if (status.error) {
-      alert(status.error);
-      return;
-    }
-
-    // redirect to next page
-    // we will need to fix this later
-    const newID = object.data.id + 1;
-    history.push(`/view/${newID}`);
+  //   const status = await deleteObjectById(object.data._id);
+  //
+  //   if (status.error) {
+  //     alert(status.error);
+  //     return;
+  //   }
+  //
+  //   // redirect to next page
+  //   // we will need to fix this later
+  //   const newID = object.data.id + 1;
+  //   history.push(`/view/${newID}`);
   }
 
   return (
     <>
       {/* show type your viewing */}
-      <ViewTypeFragment value={`Edit ${type}`} />
+      <ViewTypeFragment value={`Edit ${object.type}`} />
       <hr />
 
-      {object.data.id === 0 ? (
+      {object.id === 0 ? (
         <Box mx={4} my={2}>
           <Box>
             <h2>Pick Object Type</h2>
@@ -195,13 +197,13 @@ const EditObject: FunctionComponent = () => {
         justifyContent='space-around'
       >
         <Button
-          href={`/view/${object.data.id}`}
+          href={`/view/${object.id}`}
           style={{ padding: '10px 40px' }}
           variant='contained'
           color='primary'
           fullWidth={false}
         >
-          View {object.data.type}
+          View {object.type}
         </Button>
 
         <Button
@@ -209,21 +211,21 @@ const EditObject: FunctionComponent = () => {
           variant='contained'
           color='primary'
           fullWidth={false}
-          onClick={saveObject}
+          // onClick={saveObject}
         >
-          Save {object.data.type}
+          Save {object.type}
         </Button>
       </Box>
 
-      {object.data.id !== 0 ? (
+      {object.id !== 0 ? (
         <Box m={3}>
           <Button
             variant='contained'
             color='secondary'
             fullWidth={true}
-            onClick={deleteObject}
+            // onClick={deleteObject}
           >
-            Delete {object.data.type}
+            Delete {object.type}
           </Button>
         </Box>
       ) : null}