+++ src/view/fragments/ViewFields.tsx
@@ -2,19 +2,28 @@ import React, { FunctionComponent } from 'react';
 import ViewValueFragment from './ViewValueFragment';
 import BoxContainer from '../../components/BoxContainer';
 
-const ViewFields: FunctionComponent<any> = ({ object, template }) => {
+const ViewFields: FunctionComponent<any> = ({ object, fields }) => {
   return (
     <>
-      {template.fields.map((key: any, index: number) => (
-        <div key={index}>
-          <BoxContainer>
-            <span>{object.type} title:</span>
-            <ViewValueFragment value={object[key]} />
-          </BoxContainer>
-        </div>
-      ))}
+      {fields.map((key: any, index: number) => {
+        if (typeof object[key] === 'string') {
+          return (
+            <div key={index}>
+              <BoxContainer>
+                <span>{object.type} title:</span>
+                <ViewValueFragment value={object[key]} />
+              </BoxContainer>
+            </div>
+          )
+        }
+
+        if (typeof object[key] === 'object') {
+          // ViewObjectPartial
+          return <div>test</div>
+        }
+      })}
     </>
   );
-};
+}
 
 export default ViewFields;